+++ src/nvim/ex_getln.c
@@ -2634,13 +2634,13 @@ static void cursorcmd(void)
     return;
 
   if (cmdmsg_rl) {
-    msg_row = cmdline_row  + (ccline.cmdspos / (int)(Columns - 1));
-    msg_col = (int)Columns - (ccline.cmdspos % (int)(Columns - 1)) - 1;
+    msg_row = cmdline_row  + (ccline.cmdspos / (Columns - 1));
+    msg_col = Columns - (ccline.cmdspos % (Columns - 1)) - 1;
     if (msg_row <= 0)
       msg_row = Rows - 1;
   } else {
-    msg_row = cmdline_row + (ccline.cmdspos / (int)Columns);
-    msg_col = ccline.cmdspos % (int)Columns;
+    msg_row = cmdline_row + (ccline.cmdspos / Columns);
+    msg_col = ccline.cmdspos % Columns;
     if (msg_row >= Rows)
       msg_row = Rows - 1;
   }
@@ -3240,7 +3240,7 @@ static int showmatches(expand_T *xp, int wildmenu)
     else {
       /* compute the number of columns and lines for the listing */
       maxlen += 2;          /* two spaces between file names */
-      columns = ((int)Columns + 2) / maxlen;
+      columns = (Columns + 2) / maxlen;
       if (columns < 1)
         columns = 1;
       lines = (num_files + columns - 1) / columns;
@@ -4867,9 +4867,9 @@ void ex_history(exarg_T *eap)
           msg_putchar('\n');
           sprintf((char *)IObuff, "%c%6d  ", i == idx ? '>' : ' ',
               hist[i].hisnum);
-          if (vim_strsize(hist[i].hisstr) > (int)Columns - 10)
+          if (vim_strsize(hist[i].hisstr) > Columns - 10)
             trunc_string(hist[i].hisstr, IObuff + STRLEN(IObuff),
-                (int)Columns - 10, IOSIZE - (int)STRLEN(IObuff));
+                         Columns - 10, IOSIZE - (int)STRLEN(IObuff));
           else
             STRCAT(IObuff, hist[i].hisstr);
           msg_outtrans(IObuff);
@@ -4973,7 +4973,7 @@ static int ex_window(void)
   cmdmod.tab = 0;
 
   /* Create a window for the command-line buffer. */
-  if (win_split((int)p_cwh, WSP_BOT) == FAIL) {
+  if (win_split(p_cwh, WSP_BOT) == FAIL) {
     beep_flush();
     unblock_autocmds();
     return K_IGNORE;
